refactor: change image extension
- use .avif images where possible
