+++ scroll-animation/src/hooks/useWindowSize.ts
@@ -1,0 +1,23 @@
+import { useState, useEffect } from "react";
+
+export const useWindowSize = () => {
+  function getSize() {
+    return {
+      width: window.innerWidth,
+      height: window.innerHeight,
+    };
+  }
+
+  const [windowSize, setWindowSize] = useState(getSize());
+
+  useEffect(() => {
+    function handleResize() {
+      setWindowSize(getSize());
+    }
+
+    window.addEventListener("resize", handleResize);
+    return () => window.removeEventListener("resize", handleResize);
+  }, []);
+
+  return windowSize;
+};
